Enhanced test case for local testin
