+++ src/workflow/mod.rs
@@ -1,0 +1,3 @@
+pub mod index_destination;
+pub mod preparation;
+pub mod process_sources;
